requirements.txt
Babel>=1.3
eventlet>=0.13.0
iso8601>=0.1.9
oslo.config>=1.2.0
pbr>=0.6,!=0.7,<1.0
posix_ipc
python-glanceclient>=0.9.0
python-keystoneclient>=0.9.0
